MARKER FOR ODD BUG.
This commit is a marker for an odd bug we saw, in case
it occurs again.  That is, we got the robot code to
reboot over and over again, and it's not clear why.
It had something to do with invoking it as a static
method.  This is a comment only commit.